Fix exception raised when attempting to vote
Seems like Bukkit / Spigot API changes have caused the vote menu to
raise an exception and crash when attempting to vote in an election.

Either that, or it's because we removed cached player state from the
worlds we're using this plugin in, and we didn't notice it broke things.

In any case, a new version of triumph-gui (3.1.6 -- weirdly we were
using 3.2 previously, but such a version doesn't actually seem to exist)
coupled with some changes to make sure we don't pass a null owner to the
SkullBuilder.owne() api seems to have fixed the issues.